update
- updated module loading/unloading as a seperated module
- combined dev tools in a module
- added module loading/unloading for admin.py
- moved disaster related functions from admin.py to disasters.py
