+++ tg_bot/modules/admin.py
@@ -1,9 +1,5 @@
 import html
-import subprocess
-import importlib
-import os
 import json
-import sys
 import requests
 
 from typing import List
@@ -12,502 +8,13 @@ from time import sleep
 from telegram import Bot, Update, ParseMode
 from telegram.error import BadRequest
 from telegram.ext import CommandHandler, Filters, run_async
-from telegram.utils.helpers import escape_markdown, mention_html
+from telegram.utils.helpers import mention_html
 
-from tg_bot import dispatcher, WHITELIST_USERS, SUPPORT_USERS, SUDO_USERS, DEV_USERS, OWNER_ID, TOKEN
-from tg_bot.__main__ import IMPORTED, HELPABLE, MIGRATEABLE, STATS, USER_INFO, DATA_IMPORT, DATA_EXPORT, CHAT_SETTINGS, USER_SETTINGS 
+from tg_bot import dispatcher, TOKEN
 from tg_bot.modules.disable import DisableAbleCommandHandler
-from tg_bot.modules.log_channel import loggable, gloggable
-from tg_bot.modules.helper_funcs.chat_status import bot_admin, can_promote, user_admin, can_pin, sudo_plus, dev_plus, connection_status
+from tg_bot.modules.log_channel import loggable
+from tg_bot.modules.helper_funcs.chat_status import bot_admin, can_promote, user_admin, can_pin, connection_status
 from tg_bot.modules.helper_funcs.extraction import extract_user, extract_user_and_text
-
-
-@run_async
-@dev_plus
-@gloggable
-def addsudo(bot: Bot, update: Update, args: List[str]) -> str:
-    
-    message = update.effective_message
-    user = update.effective_user
-    chat = update.effective_chat
-
-    user_id = extract_user(message, args)
-    user_member = bot.getChat(user_id)
-    rt, log_message = "", ""
-
-    if not user_id:
-        message.reply_text("That...is a chat! baka ka omae?")
-        return log_message
-
-    if user_id == bot.id:
-        message.reply_text("This does not work that way.")
-        return log_message
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
-        data = json.load(infile)
-
-    if user_id in SUDO_USERS:
-        message.reply_text("This member is already a Dragon Disaster")
-        return log_message
-
-    if user_id in SUPPORT_USERS:
-        rt += ("This user is already a Demon Disaster, Promoting to Dragon Disaster.")
-        data['supports'].remove(user_id)
-        SUPPORT_USERS.remove(user_id)
-
-    if user_id in WHITELIST_USERS:
-        rt += ("This user is already a Wolf, Promoting to Dragon Disaster.")
-        data['whitelists'].remove(user_id)
-        WHITELIST_USERS.remove(user_id)
-
-    data['sudos'].append(user_id)
-    SUDO_USERS.append(user_id)
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
-        json.dump(data, outfile, indent=4)
-
-    update.effective_message.reply_text(rt + "\nSuccessfully set Disaster level of {} to Dragon!".format(user_member.first_name))
-    
-    if chat.type != 'private':
-        log_message += "<b>{}:</b>\n".format(html.escape(chat.title))
-
-    log_message += "#SUDO"\
-                   "\n<b>Admin:</b> {} "\
-                   "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
-                                              mention_html(user_member.id, user_member.first_name))
-
-    return log_message
-
-
-@run_async
-@dev_plus
-@gloggable
-def removesudo(bot: Bot, update: Update, args: List[str]) -> str:
-
-    message = update.effective_message
-    user = update.effective_user
-    chat = update.effective_chat
-
-    user_id = extract_user(message, args)
-    user_member = bot.getChat(user_id)
-    log_message = ""
-
-    if not user_id:
-        message.reply_text("I can't remove a chat from sudo list!")
-        return log_message
-
-    if user_id == bot.id:
-        message.reply_text("Why are you removing me from my sudo list?")
-        return log_message
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
-        data = json.load(infile)
-
-    if user_id in SUDO_USERS:
-        message.reply_text("Demoting to normal user")
-        SUDO_USERS.remove(user_id)
-        data['sudos'].remove(user_id)
-
-        with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
-            json.dump(data, outfile, indent=4)
-    
-        if chat.type != 'private':
-            log_message += "<b>{}:</b>\n".format(html.escape(chat.title))
-
-        log_message += "#UNSUDO"\
-                       "\n<b>Admin:</b> {} "\
-                       "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
-                                                  mention_html(user_member.id, user_member.first_name))
-    
-        return log_message
-
-    else:
-        message.reply_text("This user is not a Dragon Disaster!")
-        return log_message
-
-
-@run_async
-@dev_plus
-@gloggable
-def addsupport(bot: Bot, update: Update, args: List[str]) -> str:
-
-    message = update.effective_message
-    user = update.effective_user
-    chat = update.effective_chat
-
-    user_id = extract_user(message, args)
-    user_member = bot.getChat(user_id)
-    rt, log_message = "", ""
-
-    if not user_id:
-        message.reply_text("I can't add a chat to support list!")
-        return log_message
-
-    if user_id == bot.id:
-        message.reply_text("Why are you adding me to my support list?")
-        return log_message
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
-        data = json.load(infile)
-
-    if user_id in SUDO_USERS:
-        rt += ("Demoting status of this Dragon to Demon")
-        data['sudos'].remove(user_id)
-        SUDO_USERS.remove(user_id)
-
-    if user_id in SUPPORT_USERS:
-        message.reply_text("This user is already a Demon Disaster.")
-        return log_message
-
-    if user_id in WHITELIST_USERS:
-        rt+=("Promoting Disaster level from Wolf to Demon")
-        data['whitelists'].remove(user_id)
-        WHITELIST_USERS.remove(user_id)
-
-    data['supports'].append(user_id)
-    SUPPORT_USERS.append(user_id)
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
-        json.dump(data, outfile, indent=4)
-    
-    update.effective_message.reply_text(rt + "\n{} was added as a Demon Disaster!".format(user_member.first_name))
-    
-    if chat.type != 'private':
-        log_message += "<b>{}:</b>\n".format(html.escape(chat.title))
-
-    log_message += "#SUPPORT"\
-                   "\n<b>Admin:</b> {} "\
-                   "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
-                                              mention_html(user_member.id, user_member.first_name))
-    
-    return log_message
-
-
-@run_async
-@dev_plus
-@gloggable
-def removesupport(bot: Bot, update: Update, args: List[str]) -> str:
-
-    message = update.effective_message
-    user = update.effective_user
-    chat = update.effective_chat
-
-    user_id = extract_user(message, args)
-    user_member = bot.getChat(user_id)
-    log_message = ""
-
-    if not user_id:
-        message.reply_text("I can't remove a chat from support list!")
-        return log_message
-
-    if user_id == bot.id:
-        message.reply_text("Why are you removing me from my support list?")
-        return log_message
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
-        data = json.load(infile)
-
-    if user_id in SUPPORT_USERS:
-        message.reply_text("Demoting to Civilian")
-        SUPPORT_USERS.remove(user_id)
-        data['supports'].remove(user_id)
-
-        with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
-            json.dump(data, outfile, indent=4)
-    
-        if chat.type != 'private':
-            log_message += "<b>{}:</b>\n".format(html.escape(chat.title))
-
-        log_message += "#UNSUPPORT"\
-                       "\n<b>Admin:</b> {} "\
-                       "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
-                                                  mention_html(user_member.id, user_member.first_name))
-        
-        return log_message
-
-    else:
-        message.reply_text("This user is not a Demon level Disaster!")
-        return ""
-
-
-@run_async
-@dev_plus
-@gloggable
-def addwhitelist(bot: Bot, update: Update, args: List[str]) -> str:
-
-    message = update.effective_message
-    user = update.effective_user
-    chat = update.effective_chat
-
-    user_id = extract_user(message, args)
-    user_member = bot.getChat(user_id)
-    rt, log_message = "", ""
-
-    if not user_id:
-        message.reply_text("I can't add a chat to whitelist list!")
-        return log_message
-
-    if user_id == bot.id:
-        message.reply_text("Why are you adding me to my whitelist list?")
-        return log_message
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
-        data = json.load(infile)
-
-    if user_id in SUDO_USERS:
-        rt += ("This member is a Dragon Disaster, Demoting to Wolf.")
-        data['sudos'].remove(user_id)
-        SUDO_USERS.remove(user_id)
-
-    if user_id in SUPPORT_USERS:
-        rt += ("This user is already a Demon Disaster, Demoting to Wolf.")
-        data['supports'].remove(user_id)
-        SUPPORT_USERS.remove(user_id)
-
-    if user_id in WHITELIST_USERS:
-        message.reply_text("This user is already a Wolf Disaster.")
-        return log_message
-
-    data['whitelists'].append(user_id)
-    WHITELIST_USERS.append(user_id)
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
-        json.dump(data, outfile, indent=4)
-
-    update.effective_message.reply_text(rt + "\nSuccessfully promoted {} to a Wolf Disaster!".format(user_member.first_name))
-    
-    if chat.type != 'private':
-        log_message += "<b>{}:</b>\n".format(html.escape(chat.title))
-
-    log_message += "#WHITELIST"\
-                   "\n<b>Admin:</b> {} "\
-                   "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
-                                              mention_html(user_member.id, user_member.first_name))
-    
-    return log_message
-
-
-@run_async
-@dev_plus
-@gloggable
-def removewhitelist(bot: Bot, update: Update, args: List[str]) -> str:
-
-    message = update.effective_message
-    user = update.effective_user
-    chat = update.effective_chat
-
-    user_id = extract_user(message, args)
-    user_member = bot.getChat(user_id)
-    log_message = ""
-
-    if not user_id:
-        message.reply_text("I can't remove a chat from whitelist list!")
-        return ""
-
-    if user_id == bot.id:
-        message.reply_text("Why are you removing me from my whitelist list?")
-        return log_message
-
-    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
-        data = json.load(infile)
-
-    if user_id in WHITELIST_USERS:
-        message.reply_text("Demoting to normal user")
-        WHITELIST_USERS.remove(user_id)
-        data['whitelists'].remove(user_id)
-
-        with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
-            json.dump(data, outfile, indent=4)
-    
-        if chat.type != 'private':
-            log_message += "<b>{}:</b>\n".format(html.escape(chat.title))
-
-        log_message += "#UNWHITELIST"\
-                       "\n<b>Admin:</b> {} "\
-                       "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
-                                                  mention_html(user_member.id, user_member.first_name))
-        
-        return log_message
-    else:
-        message.reply_text("This user is not a Wolf Disaster!")
-        return ""
-
-
-@run_async
-@dev_plus
-def gitpull(bot: Bot, update: Update):
-
-    sent_msg = update.effective_message.reply_text("Pulling all changes from remote and then attempting to restart.")
-    subprocess.Popen('git pull', stdout=subprocess.PIPE, shell=True)
-
-    sent_msg_text = sent_msg.text + "\n\nChanges pulled...I guess.. Restarting in "
-
-    for i in reversed(range(5)):
-        sent_msg.edit_text(sent_msg_text + str(i + 1))
-        sleep(1)
-    
-    sent_msg.edit_text("Restarted.")
-    
-    os.system('restart.bat')
-    os.execv('start.bat', sys.argv)
-
-
-@run_async
-@dev_plus
-def restart(bot: Bot, update: Update):
-
-    update.effective_message.reply_text("Starting a new instance and shutting down this one")
-
-    os.system('restart.bat')
-    os.execv('start.bat', sys.argv)
-
-
-@run_async
-@dev_plus
-def load(bot: Bot, update: Update):
-
-    message = update.effective_message
-    text = message.text.split(" ", 1)[1]
-    load_messasge = message.reply_text(f"Attempting to load module : <b>{text}</b>", parse_mode=ParseMode.HTML)
-    
-    try:
-        imported_module = importlib.import_module("tg_bot.modules." + text)
-    except:
-        load_messasge.edit_text("Does that module even exist?")
-        return
-
-    if not hasattr(imported_module, "__mod_name__"):
-        imported_module.__mod_name__ = imported_module.__name__
-
-    if not imported_module.__mod_name__.lower() in IMPORTED:
-        IMPORTED[imported_module.__mod_name__.lower()] = imported_module
-    else:
-        load_messasge.edit_text("Module already loaded.")
-        return
-    
-    if "__handlers__" in dir(imported_module):
-        handlers = imported_module.__handlers__
-        for handler in handlers:
-            if type(handler) != tuple:
-                dispatcher.add_handler(handler)
-            else:
-                handler_name, priority = handler
-                dispatcher.add_handler(handler_name, priority)
-    else:
-        IMPORTED.pop(imported_module.__mod_name__.lower())
-        load_messasge.edit_text("The module doesn't have a handler list specified!")
-        return
-
-    if hasattr(imported_module, "__help__") and imported_module.__help__:
-        HELPABLE[imported_module.__mod_name__.lower()] = imported_module
-
-    # Chats to migrate on chat_migrated events
-    if hasattr(imported_module, "__migrate__"):
-        MIGRATEABLE.append(imported_module)
-
-    if hasattr(imported_module, "__stats__"):
-        STATS.append(imported_module)
-
-    if hasattr(imported_module, "__user_info__"):
-        USER_INFO.append(imported_module)
-
-    if hasattr(imported_module, "__import_data__"):
-        DATA_IMPORT.append(imported_module)
-
-    if hasattr(imported_module, "__export_data__"):
-        DATA_EXPORT.append(imported_module)
-
-    if hasattr(imported_module, "__chat_settings__"):
-        CHAT_SETTINGS[imported_module.__mod_name__.lower()] = imported_module
-
-    if hasattr(imported_module, "__user_settings__"):
-        USER_SETTINGS[imported_module.__mod_name__.lower()] = imported_module
-
-    load_messasge.edit_text("Successfully loaded module : <b>{}</b>".format(text), parse_mode=ParseMode.HTML)
-
-
-@run_async
-@dev_plus
-def unload(bot: Bot, update: Update):
-
-    message = update.effective_message
-    text = message.text.split(" ", 1)[1]
-    unload_messasge = message.reply_text(f"Attempting to unload module : <b>{text}</b>", parse_mode=ParseMode.HTML)
-
-    try:
-        imported_module = importlib.import_module("tg_bot.modules." + text)
-    except:
-        unload_messasge.edit_text("Does that module even exist?")
-        return
-
-    if not hasattr(imported_module, "__mod_name__"):
-        imported_module.__mod_name__ = imported_module.__name__
-    
-    if imported_module.__mod_name__.lower() in IMPORTED:
-        IMPORTED.pop(imported_module.__mod_name__.lower())
-    else:
-        unload_messasge.edit_text("Can't unload something that isn't loaded.")
-        return
-    
-    if "__handlers__" in dir(imported_module):
-        handlers = imported_module.__handlers__
-        for handler in handlers:
-            if type(handler) == bool:
-                unload_messasge.edit_text("This module can't be unloaded!")
-                return
-            elif type(handler) != tuple:
-                dispatcher.remove_handler(handler)
-            else:
-                handler_name, priority = handler
-                dispatcher.remove_handler(handler_name, priority)
-    else:
-        unload_messasge.edit_text("The module doesn't have a handler list specified!")
-        return
-
-    if hasattr(imported_module, "__help__") and imported_module.__help__:
-        HELPABLE.pop(imported_module.__mod_name__.lower())
-
-    # Chats to migrate on chat_migrated events
-    if hasattr(imported_module, "__migrate__"):
-        MIGRATEABLE.remove(imported_module)
-
-    if hasattr(imported_module, "__stats__"):
-        STATS.remove(imported_module)
-
-    if hasattr(imported_module, "__user_info__"):
-        USER_INFO.remove(imported_module)
-
-    if hasattr(imported_module, "__import_data__"):
-        DATA_IMPORT.remove(imported_module)
-
-    if hasattr(imported_module, "__export_data__"):
-        DATA_EXPORT.remove(imported_module)
-
-    if hasattr(imported_module, "__chat_settings__"):
-        CHAT_SETTINGS.pop(imported_module.__mod_name__.lower())
-
-    if hasattr(imported_module, "__user_settings__"):
-        USER_SETTINGS.pop(imported_module.__mod_name__.lower())
-
-    unload_messasge.edit_text(f"Successfully unloaded module : <b>{text}</b>", parse_mode=ParseMode.HTML)
-
-
-@run_async
-@connection_status
-@sudo_plus
-def listmodules(bot: Bot, update: Update):
-
-    message = update.effective_message
-    module_list = []
-
-    for helpable_module in HELPABLE:
-        helpable_module_info = IMPORTED[helpable_module]
-        file_info = IMPORTED[helpable_module_info.__mod_name__.lower()]
-        file_name = file_info.__name__.rsplit("tg_bot.modules.", 1)[1]
-        mod_name = file_info.__mod_name__
-        module_list.append(f'- <code>{mod_name} ({file_name})</code>\n')
-    module_list = "Following modules are loaded : \n\n" + ''.join(module_list)
-    message.reply_text(module_list, parse_mode=ParseMode.HTML)
 
 
 @run_async
@@ -745,6 +252,7 @@ def unpin(bot: Bot, update: Update) -> str:
 
     return log_message
 
+
 @run_async
 @bot_admin
 @user_admin
@@ -763,6 +271,7 @@ def invite(bot: Bot, update: Update):
             update.effective_message.reply_text("I don't have access to the invite link, try changing my permissions!")
     else:
         update.effective_message.reply_text("I can only give you invite links for supergroups and channels, sorry!")
+
 
 @run_async
 @connection_status
@@ -820,20 +329,6 @@ DEMOTE_HANDLER = CommandHandler("demote", demote, pass_args=True)
 
 SET_TITLE_HANDLER = CommandHandler("settitle", set_title, pass_args=True)
 
-GITPULL_HANDLER = CommandHandler("gitpull", gitpull)
-RESTART_HANDLER = CommandHandler("restart", restart)
-
-LOAD_HANDLER = CommandHandler("load", load)
-UNLOAD_HANDLER = CommandHandler("unload", unload)
-LISTMODULES_HANDLER = CommandHandler("listmodules", listmodules)
-
-SUDO_HANDLER = CommandHandler(("addsudo", "adddragon"), addsudo, pass_args=True)
-UNSUDO_HANDLER = CommandHandler(("removesudo", "removedragon"), removesudo, pass_args=True)
-SUPPORT_HANDLER = CommandHandler(("addsupport", "adddemon"), addsupport, pass_args=True)
-UNSUPPORT_HANDLER = CommandHandler(("removesupport", "removedemon"), removesupport, pass_args=True)
-WHITELIST_HANDLER = CommandHandler(("addwhitelist", "addwolf"), addwhitelist, pass_args=True)
-UNWHITELIST_HANDLER = CommandHandler(("removewhitelist", "removewolf"), removewhitelist, pass_args=True)
-
 dispatcher.add_handler(ADMINLIST_HANDLER)
 dispatcher.add_handler(PIN_HANDLER)
 dispatcher.add_handler(UNPIN_HANDLER)
@@ -841,16 +336,8 @@ dispatcher.add_handler(INVITE_HANDLER)
 dispatcher.add_handler(PROMOTE_HANDLER)
 dispatcher.add_handler(DEMOTE_HANDLER)
 dispatcher.add_handler(SET_TITLE_HANDLER)
-dispatcher.add_handler(GITPULL_HANDLER)
-dispatcher.add_handler(RESTART_HANDLER)
-dispatcher.add_handler(LOAD_HANDLER)
-dispatcher.add_handler(UNLOAD_HANDLER)
-dispatcher.add_handler(LISTMODULES_HANDLER)
-dispatcher.add_handler(SUDO_HANDLER)
-dispatcher.add_handler(UNSUDO_HANDLER)
-dispatcher.add_handler(SUPPORT_HANDLER)
-dispatcher.add_handler(UNSUPPORT_HANDLER)
-dispatcher.add_handler(WHITELIST_HANDLER)
-dispatcher.add_handler(UNWHITELIST_HANDLER)
 
 __mod_name__ = "Admin"
+__command_list__ = ["adminlist","admins", "invitelink"]
+__handlers__ = [ADMINLIST_HANDLER, PIN_HANDLER, UNPIN_HANDLER,
+                INVITE_HANDLER, PROMOTE_HANDLER, DEMOTE_HANDLER, SET_TITLE_HANDLER]
