+++ tg_bot/modules/disasters.py
@@ -1,16 +1,323 @@
 import json
+import html
 import os
 
-from telegram import Bot, Update, ParseMode
-from telegram.ext import run_async
+from typing import List, Optional
 
-from tg_bot import dispatcher, SUDO_USERS, DEV_USERS, WHITELIST_USERS, OWNER_ID, SUPPORT_USERS
-from tg_bot.modules.disable import DisableAbleCommandHandler
-from tg_bot.modules.helper_funcs.chat_status import whitelist_plus
+from telegram import Bot, Update, ParseMode, TelegramError
+from telegram.ext import CommandHandler, run_async
+from telegram.utils.helpers import mention_html
 
-def get_user_list(config, key):
-    with open('{}/tg_bot/{}'.format(os.getcwd(), config), 'r') as json_file:
-        return json.load(json_file)[key]
+from tg_bot import dispatcher, WHITELIST_USERS, SUPPORT_USERS, SUDO_USERS, DEV_USERS, OWNER_ID
+from tg_bot.modules.log_channel import gloggable
+from tg_bot.modules.helper_funcs.chat_status import whitelist_plus, dev_plus
+from tg_bot.modules.helper_funcs.extraction import extract_user
+
+
+def check_user_id(user_id: int, bot: Bot) -> Optional[str]:
+
+    if not user_id:
+        reply = "That...is a chat! baka ka omae?"
+    
+    elif user_id == bot.id:
+        reply = "This does not work that way."
+
+    else:
+        reply = None
+    
+    return reply
+
+
+@run_async
+@dev_plus
+@gloggable
+def addsudo(bot: Bot, update: Update, args: List[str]) -> str:
+    
+    message = update.effective_message
+    user = update.effective_user
+    chat = update.effective_chat
+
+    user_id = extract_user(message, args)
+    user_member = bot.getChat(user_id)
+    rt = ""
+
+    reply = check_user_id(user_id, bot)
+    if reply:
+        message.reply_text(reply)
+        return ""
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
+        data = json.load(infile)
+
+    if user_id in SUDO_USERS:
+        message.reply_text("This member is already a Dragon Disaster")
+        return ""
+
+    if user_id in SUPPORT_USERS:
+        rt += ("This user is already a Demon Disaster, Promoting to Dragon Disaster.")
+        data['supports'].remove(user_id)
+        SUPPORT_USERS.remove(user_id)
+
+    if user_id in WHITELIST_USERS:
+        rt += ("This user is already a Wolf, Promoting to Dragon Disaster.")
+        data['whitelists'].remove(user_id)
+        WHITELIST_USERS.remove(user_id)
+
+    data['sudos'].append(user_id)
+    SUDO_USERS.append(user_id)
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
+        json.dump(data, outfile, indent=4)
+
+    update.effective_message.reply_text(rt + "\nSuccessfully set Disaster level of {} to Dragon!".format(user_member.first_name))
+
+    log_message = "#SUDO"\
+                  "\n<b>Admin:</b> {} "\
+                  "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
+                                             mention_html(user_member.id, user_member.first_name))
+    
+    if chat.type != 'private':
+        log_message = "<b>{}:</b>\n".format(html.escape(chat.title)) + log_message
+
+    return log_message
+
+
+@run_async
+@dev_plus
+@gloggable
+def addsupport(bot: Bot, update: Update, args: List[str]) -> str:
+
+    message = update.effective_message
+    user = update.effective_user
+    chat = update.effective_chat
+
+    user_id = extract_user(message, args)
+    user_member = bot.getChat(user_id)
+    rt = ""
+
+    reply = check_user_id(user_id, bot)
+    if reply:
+        message.reply_text(reply)
+        return ""
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
+        data = json.load(infile)
+
+    if user_id in SUDO_USERS:
+        rt += ("Demoting status of this Dragon to Demon")
+        data['sudos'].remove(user_id)
+        SUDO_USERS.remove(user_id)
+
+    if user_id in SUPPORT_USERS:
+        message.reply_text("This user is already a Demon Disaster.")
+        return ""
+
+    if user_id in WHITELIST_USERS:
+        rt+=("Promoting Disaster level from Wolf to Demon")
+        data['whitelists'].remove(user_id)
+        WHITELIST_USERS.remove(user_id)
+
+    data['supports'].append(user_id)
+    SUPPORT_USERS.append(user_id)
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
+        json.dump(data, outfile, indent=4)
+    
+    update.effective_message.reply_text(rt + "\n{} was added as a Demon Disaster!".format(user_member.first_name))
+
+    log_message = "#SUPPORT"\
+                  "\n<b>Admin:</b> {} "\
+                  "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
+                                             mention_html(user_member.id, user_member.first_name))
+    
+    if chat.type != 'private':
+        log_message = "<b>{}:</b>\n".format(html.escape(chat.title)) + log_message
+    
+    return log_message
+
+
+@run_async
+@dev_plus
+@gloggable
+def addwhitelist(bot: Bot, update: Update, args: List[str]) -> str:
+
+    message = update.effective_message
+    user = update.effective_user
+    chat = update.effective_chat
+
+    user_id = extract_user(message, args)
+    user_member = bot.getChat(user_id)
+    rt = ""
+
+    reply = check_user_id(user_id, bot)
+    if reply:
+        message.reply_text(reply)
+        return ""
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
+        data = json.load(infile)
+
+    if user_id in SUDO_USERS:
+        rt += ("This member is a Dragon Disaster, Demoting to Wolf.")
+        data['sudos'].remove(user_id)
+        SUDO_USERS.remove(user_id)
+
+    if user_id in SUPPORT_USERS:
+        rt += ("This user is already a Demon Disaster, Demoting to Wolf.")
+        data['supports'].remove(user_id)
+        SUPPORT_USERS.remove(user_id)
+
+    if user_id in WHITELIST_USERS:
+        message.reply_text("This user is already a Wolf Disaster.")
+        return ""
+
+    data['whitelists'].append(user_id)
+    WHITELIST_USERS.append(user_id)
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
+        json.dump(data, outfile, indent=4)
+
+    update.effective_message.reply_text(rt + "\nSuccessfully promoted {} to a Wolf Disaster!".format(user_member.first_name))
+
+    log_message = "#WHITELIST"\
+                  "\n<b>Admin:</b> {} "\
+                  "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
+                                             mention_html(user_member.id, user_member.first_name))
+    
+    if chat.type != 'private':
+        log_message = "<b>{}:</b>\n".format(html.escape(chat.title)) + log_message
+    
+    return log_message
+
+
+@run_async
+@dev_plus
+@gloggable
+def removesudo(bot: Bot, update: Update, args: List[str]) -> str:
+
+    message = update.effective_message
+    user = update.effective_user
+    chat = update.effective_chat
+
+    user_id = extract_user(message, args)
+    user_member = bot.getChat(user_id)
+
+    reply = check_user_id(user_id, bot)
+    if reply:
+        message.reply_text(reply)
+        return ""
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
+        data = json.load(infile)
+
+    if user_id in SUDO_USERS:
+        message.reply_text("Demoting to normal user")
+        SUDO_USERS.remove(user_id)
+        data['sudos'].remove(user_id)
+
+        with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
+            json.dump(data, outfile, indent=4)
+
+        log_message = "#UNSUDO"\
+                      "\n<b>Admin:</b> {} "\
+                      "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
+                                                 mention_html(user_member.id, user_member.first_name))
+    
+        if chat.type != 'private':
+            log_message = "<b>{}:</b>\n".format(html.escape(chat.title)) + log_message
+    
+        return log_message
+
+    else:
+        message.reply_text("This user is not a Dragon Disaster!")
+        return ""
+
+
+
+@run_async
+@dev_plus
+@gloggable
+def removesupport(bot: Bot, update: Update, args: List[str]) -> str:
+
+    message = update.effective_message
+    user = update.effective_user
+    chat = update.effective_chat
+
+    user_id = extract_user(message, args)
+    user_member = bot.getChat(user_id)
+
+    reply = check_user_id(user_id, bot)
+    if reply:
+        message.reply_text(reply)
+        return ""
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
+        data = json.load(infile)
+
+    if user_id in SUPPORT_USERS:
+        message.reply_text("Demoting to Civilian")
+        SUPPORT_USERS.remove(user_id)
+        data['supports'].remove(user_id)
+
+        with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
+            json.dump(data, outfile, indent=4)
+
+        log_message = "#UNSUPPORT"\
+                      "\n<b>Admin:</b> {} "\
+                      "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
+                                                 mention_html(user_member.id, user_member.first_name))
+    
+        if chat.type != 'private':
+            log_message = "<b>{}:</b>\n".format(html.escape(chat.title)) + log_message
+        
+        return log_message
+
+    else:
+        message.reply_text("This user is not a Demon level Disaster!")
+        return ""
+
+
+@run_async
+@dev_plus
+@gloggable
+def removewhitelist(bot: Bot, update: Update, args: List[str]) -> str:
+
+    message = update.effective_message
+    user = update.effective_user
+    chat = update.effective_chat
+
+    user_id = extract_user(message, args)
+    user_member = bot.getChat(user_id)
+
+    reply = check_user_id(user_id, bot)
+    if reply:
+        message.reply_text(reply)
+        return ""
+
+    with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'r') as infile:
+        data = json.load(infile)
+
+    if user_id in WHITELIST_USERS:
+        message.reply_text("Demoting to normal user")
+        WHITELIST_USERS.remove(user_id)
+        data['whitelists'].remove(user_id)
+
+        with open('{}/tg_bot/elevated_users.json'.format(os.getcwd()), 'w') as outfile:
+            json.dump(data, outfile, indent=4)
+
+        log_message = "#UNWHITELIST"\
+                      "\n<b>Admin:</b> {} "\
+                      "\n<b>User:</b> {}".format(mention_html(user.id, user.first_name),
+                                                 mention_html(user_member.id, user_member.first_name))
+    
+        if chat.type != 'private':
+            log_message = "<b>{}:</b>\n".format(html.escape(chat.title)) + log_message
+        
+        return log_message
+    else:
+        message.reply_text("This user is not a Wolf Disaster!")
+        return ""
+
 
 @run_async
 @whitelist_plus
@@ -23,7 +330,7 @@ def whitelistlist(bot: Bot, update: Update):
             user = bot.get_chat(user_id)
             first_name = user.first_name
             reply += """• <a href="tg://user?id={}">{}</a>\n""".format(user_id, first_name)
-        except:
+        except TelegramError:
             pass
     update.effective_message.reply_text(reply, parse_mode=ParseMode.HTML)
 
@@ -39,10 +346,9 @@ def supportlist(bot: Bot, update: Update):
             user = bot.get_chat(user_id)
             first_name = user.first_name
             reply += """• <a href="tg://user?id={}">{}</a>\n""".format(user_id, first_name)
-        except:
+        except TelegramError:
             pass
     update.effective_message.reply_text(reply, parse_mode=ParseMode.HTML)
-
 
 
 @run_async
@@ -57,26 +363,24 @@ def sudolist(bot: Bot, update: Update):
             user = bot.get_chat(user_id)
             first_name = user.first_name
             reply += """• <a href="tg://user?id={}">{}</a>\n""".format(user_id, first_name)
-        except:
+        except TelegramError:
             pass
     update.effective_message.reply_text(reply, parse_mode=ParseMode.HTML)
-
 
 
 @run_async
 @whitelist_plus
 def devlist(bot: Bot, update: Update):
 
+    true_dev = list(set(DEV_USERS) - set([OWNER_ID]))
     reply = "<b>Hero Association Members ⚡️:</b>\n"
-    for each_user in DEV_USERS:
+    for each_user in true_dev:
         user_id = int(each_user)
-        if user_id == int(OWNER_ID):
-            continue
         try:
             user = bot.get_chat(user_id)
             first_name = user.first_name
             reply += """• <a href="tg://user?id={}">{}</a>\n""".format(user_id, first_name)
-        except:
+        except TelegramError:
             pass
     update.effective_message.reply_text(reply, parse_mode=ParseMode.HTML)
 
@@ -88,10 +392,25 @@ __help__ = """
  - /devlist - List dev users.
 """
 
-WHITELISTLIST_HANDLER = DisableAbleCommandHandler(["whitelistlist", "wolves"], whitelistlist)
-SUPPORTLIST_HANDLER = DisableAbleCommandHandler(["supportlist", "demons"], supportlist)
-SUDOLIST_HANDLER = DisableAbleCommandHandler(["sudolist", "dragons"], sudolist)
-DEVLIST_HANDLER = DisableAbleCommandHandler(["devlist", "heroes"], devlist)
+
+SUDO_HANDLER = CommandHandler(("addsudo", "adddragon"), addsudo, pass_args=True)
+SUPPORT_HANDLER = CommandHandler(("addsupport", "adddemon"), addsupport, pass_args=True)
+WHITELIST_HANDLER = CommandHandler(("addwhitelist", "addwolf"), addwhitelist, pass_args=True)
+UNSUDO_HANDLER = CommandHandler(("removesudo", "removedragon"), removesudo, pass_args=True)
+UNSUPPORT_HANDLER = CommandHandler(("removesupport", "removedemon"), removesupport, pass_args=True)
+UNWHITELIST_HANDLER = CommandHandler(("removewhitelist", "removewolf"), removewhitelist, pass_args=True)
+
+WHITELISTLIST_HANDLER = CommandHandler(["whitelistlist", "wolves"], whitelistlist)
+SUPPORTLIST_HANDLER = CommandHandler(["supportlist", "demons"], supportlist)
+SUDOLIST_HANDLER = CommandHandler(["sudolist", "dragons"], sudolist)
+DEVLIST_HANDLER = CommandHandler(["devlist", "heroes"], devlist)
+
+dispatcher.add_handler(SUDO_HANDLER)
+dispatcher.add_handler(SUPPORT_HANDLER)
+dispatcher.add_handler(WHITELIST_HANDLER)
+dispatcher.add_handler(UNSUDO_HANDLER)
+dispatcher.add_handler(UNSUPPORT_HANDLER)
+dispatcher.add_handler(UNWHITELIST_HANDLER)
 
 dispatcher.add_handler(WHITELISTLIST_HANDLER)
 dispatcher.add_handler(SUPPORTLIST_HANDLER)
@@ -99,5 +418,6 @@ dispatcher.add_handler(SUDOLIST_HANDLER)
 dispatcher.add_handler(DEVLIST_HANDLER)
 
 __mod_name__ = "Disasters"
-__command_list__ = ["whitelistlist", "wolves", "supportlist", "demons", "sudolist", "dragons", "devlist", "heroes"]
-__handlers__ = [WHITELISTLIST_HANDLER, SUPPORTLIST_HANDLER, SUDOLIST_HANDLER, DEVLIST_HANDLER]
+__handlers__ = [SUDO_HANDLER, SUPPORT_HANDLER, WHITELIST_HANDLER, 
+                UNSUDO_HANDLER, UNSUPPORT_HANDLER, UNWHITELIST_HANDLER, 
+                WHITELISTLIST_HANDLER, SUPPORTLIST_HANDLER, SUDOLIST_HANDLER, DEVLIST_HANDLER]
